EventumApp Backend V1

+++ src/controllers/events.deleteFavEvent.controller.ts
@@ -7,7 +7,7 @@ export const deleteFavEventController = async (
   res: Response,
   next: NextFunction
 ) => {
-  const favEvents = await FavEvent.deleteOne({ event_id: req.query.event_id });
+  const favEvents = await FavEvent.deleteOne({ event_id: req.query.event_id , user_id: req.user.user._id});
 
   res.status(200).json({ favEvents });
 };
